refactor(order-info): renamed order-info component file
