+++ src/pages/order/index.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 import { getIngredients } from '../../services/slices/burger-ingredients-slice';
 import styles from './order.module.css';
-import OrderInfo from '../../components/order-info/order-info';
+import OrderInfo from '../../components/order-info';
 import { useDispatch } from '../../services/hooks';
 
 export default function OrderPage() {
